Removed AdminRepository and added TestRepository.
